Add user to groups by affiliation attribute
If the affiliation field is present, create adobe connect groups for
each affiliation if needed and add the user to each group.

This doesn't remove a user from any groups if their affiliation changes.
A feature to add might be to remove the user for any affiliation groups
before adding.